get_private_data/requirements.txt
anyio==3.6.1
bidict==0.22.0
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
Flask==2.2.2
Flask-SocketIO==5.3.1
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
oauthlib==3.2.1
pandas==1.4.4
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tabulate==0.8.10
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
Werkzeug==2.2.2
zipp==3.8.1
